🩹 [fix] 자산 API 관련 파일 병합
